Move logic to skip creating bookmark.
